Move depiction/thumbnail code into model concern
